Move tr_msg(_ext) to expr.c/h.
