+++ src/expr.c
@@ -1915,3 +1915,254 @@ int parse_string_expression(struct world *mzx_world, char **_expression,
 #endif // CONFIG_DEBYTECODE
 
 #endif // outer CONFIG_DEBYTECODE
+
+// Translates message at target to the given buffer, returning location
+// of this buffer. && becomes &, &INPUT& becomes the last input string,
+// and &COUNTER& becomes the value of COUNTER. The size of the string is
+// clipped to 512 chars.
+
+#ifdef CONFIG_DEBYTECODE
+
+char *tr_msg_ext(struct world *mzx_world, char *mesg, int id, char *buffer,
+ char terminating_char)
+{
+  struct board *src_board = mzx_world->current_board;
+  char *src_ptr = mesg;
+  char current_char = *src_ptr;
+
+  int dest_pos = 0;
+  int expr_length;
+  int val, error;
+
+  while((current_char != terminating_char) && (dest_pos < ROBOT_MAX_TR-1))
+  {
+    switch(current_char)
+    {
+      case '\\':
+        if((src_ptr[1] == '(') || (src_ptr[1] == '<') || (src_ptr[1] == '>') ||
+         (src_ptr[1] == '\\') || (src_ptr[1] == terminating_char))
+        {
+          buffer[dest_pos] = src_ptr[1];
+          src_ptr += 2;
+        }
+        else
+        {
+          buffer[dest_pos] = current_char;
+          src_ptr++;
+        }
+        dest_pos++;
+        break;
+
+      case '(':
+      {
+        char number_buffer[16];
+        src_ptr++;
+        val = parse_expression(mzx_world, &src_ptr, &error, id);
+
+        expr_length = sprintf(number_buffer, "%d", val);
+        if(expr_length >= ROBOT_MAX_TR - dest_pos)
+          expr_length = ROBOT_MAX_TR - dest_pos - 1;
+
+        memcpy(buffer + dest_pos, number_buffer, expr_length);
+        dest_pos += expr_length;
+        break;
+      }
+
+      case '<':
+      {
+        src_ptr++;
+        expr_length = parse_string_expression(mzx_world, &src_ptr, id,
+         buffer + dest_pos, ROBOT_MAX_TR - dest_pos - 1);
+        dest_pos += expr_length;
+        break;
+      }
+
+      default:
+        buffer[dest_pos] = current_char;
+        src_ptr++;
+        dest_pos++;
+    }
+
+    current_char = *src_ptr;
+  }
+
+  buffer[dest_pos] = 0;
+  return src_ptr + 1;
+}
+
+#else /* !CONFIG_DEBYTECODE */
+
+char *tr_msg_ext(struct world *mzx_world, char *mesg, int id, char *buffer,
+ char terminating_char)
+{
+  struct board *src_board = mzx_world->current_board;
+  char name_buffer[256];
+  char number_buffer[16];
+  char *name_ptr;
+  char current_char;
+  char *src_ptr = mesg;
+  char *old_ptr;
+
+  size_t dest_pos = 0;
+  size_t name_length;
+  int error;
+  int val;
+
+  do
+  {
+    current_char = *src_ptr;
+
+    if((current_char == '(') && (mzx_world->version >= V268))
+    {
+      src_ptr++;
+      old_ptr = src_ptr;
+
+      val = parse_expression(mzx_world, &src_ptr, &error, id);
+      if(!error)
+      {
+        sprintf(number_buffer, "%d", val);
+        strcpy(buffer + dest_pos, number_buffer);
+        dest_pos += strlen(number_buffer);
+      }
+      else
+      {
+        buffer[dest_pos] = '(';
+        dest_pos++;
+        src_ptr = old_ptr;
+      }
+
+      current_char = *src_ptr;
+    }
+    else
+
+    if(current_char == '&')
+    {
+      src_ptr++;
+      current_char = *src_ptr;
+
+      if(current_char == '&')
+      {
+        src_ptr++;
+        buffer[dest_pos] = '&';
+        dest_pos++;
+      }
+      else
+      {
+        // Input or Counter?
+        name_ptr = name_buffer;
+
+        while(current_char)
+        {
+          if(current_char == '(' && (mzx_world->version >= V268))
+          {
+            src_ptr++;
+            val = parse_expression(mzx_world, &src_ptr, &error, id);
+            if(!error)
+            {
+              sprintf(number_buffer, "%d", val);
+              strcpy(name_ptr, number_buffer);
+              name_ptr += strlen(number_buffer);
+            }
+          }
+          else
+          {
+            *name_ptr = *src_ptr;
+            name_ptr++;
+            src_ptr++;
+          }
+
+          current_char = *src_ptr;
+
+          if(current_char == '&')
+          {
+            src_ptr++;
+            current_char = *src_ptr;
+            break;
+          }
+        }
+
+        *name_ptr = 0;
+
+        if(!memcasecmp(name_buffer, "INPUT", 6))
+        {
+          // Input
+          const char *input_string = src_board->input_string ? src_board->input_string : "";
+          name_length = strlen(input_string);
+          if(dest_pos + name_length >= ROBOT_MAX_TR)
+            name_length = ROBOT_MAX_TR - dest_pos - 1;
+
+          memcpy(buffer + dest_pos, input_string, name_length);
+          dest_pos += name_length;
+        }
+        else
+
+        // Counter or string
+        if(is_string(name_buffer))
+        {
+          // Write the value of the counter name
+          struct string str_src;
+
+          get_string(mzx_world, name_buffer, &str_src, 0);
+
+          name_length = str_src.length;
+
+          if(dest_pos + name_length >= ROBOT_MAX_TR)
+            name_length = ROBOT_MAX_TR - dest_pos - 1;
+
+          memcpy(buffer + dest_pos, str_src.value, name_length);
+          dest_pos += name_length;
+        }
+        else
+
+        // #(counter) is a hex representation.
+        if(name_buffer[0] == '+')
+        {
+          char *src = tr_int_to_hex_string(number_buffer,
+           get_counter(mzx_world, name_buffer + 1, id), &name_length);
+
+          if(dest_pos + name_length >= ROBOT_MAX_TR)
+            name_length = ROBOT_MAX_TR - dest_pos - 1;
+
+          memcpy(buffer + dest_pos, src, name_length);
+          dest_pos += name_length;
+        }
+        else
+
+        if(name_buffer[0] == '#')
+        {
+          name_length = 2;
+          if(dest_pos + name_length >= ROBOT_MAX_TR)
+            name_length = ROBOT_MAX_TR - dest_pos - 1;
+
+          sprintf(number_buffer, "%02x",
+           get_counter(mzx_world, name_buffer + 1, id));
+
+          memcpy(buffer + dest_pos, number_buffer, name_length);
+          dest_pos += name_length;
+        }
+        else
+        {
+          char *src = tr_int_to_string(number_buffer,
+           get_counter(mzx_world, name_buffer, id), &name_length);
+
+          if(dest_pos + name_length >= ROBOT_MAX_TR)
+            name_length = ROBOT_MAX_TR - dest_pos - 1;
+
+          memcpy(buffer + dest_pos, src, name_length);
+          dest_pos += name_length;
+        }
+      }
+    }
+    else
+    {
+      buffer[dest_pos] = current_char;
+      src_ptr++;
+      dest_pos++;
+    }
+  } while(current_char && (dest_pos < ROBOT_MAX_TR - 1));
+
+  buffer[dest_pos] = 0;
+  return buffer;
+}
+
+#endif /* !CONFIG_DEBYTECODE */
