+++ src/expr.h
@@ -36,6 +36,9 @@ int parse_string_expression(struct world *mzx_world, char **_expression,
  int id, char *output, size_t output_left);
 #endif
 
+CORE_LIBSPEC char *tr_msg_ext(struct world *mzx_world, char *mesg, int id,
+ char *buffer, char terminating_char);
+
 /* Guard against division by 0 and dividing INT_MIN by -1. */
 static inline int safe_divide_32(int a, int b)
 {
@@ -131,6 +134,12 @@ static inline char *tr_int_to_hex_string(char dest[9], int value, size_t *len)
 
   *len = dest + 8 - pos;
   return pos;
+}
+
+static inline char *tr_msg(struct world *mzx_world, char *mesg, int id,
+ char *buffer)
+{
+  return tr_msg_ext(mzx_world, mesg, id, buffer, 0);
 }
 
 __M_END_DECLS
